Split base classes in separate modules
Note that some methods on the SectionCalculator base class that were not general are removed.
